Create IFTTTComputerCommands.ahk
Always running script that scans a file located in Dropbox. This file contains a command for AHK to run. Once the file is created, it's read, deleted, and then the command is carried out. 

Used in conjunction with Google Assistant and IFTTT to create the files in dropbox to control the computer. 
Command descriptions, keywords and responses are provided with each command. Both in a review at the top of the file, and directly above each command.