Single Monitor Version
Added a simplified version of the Change Audio Devices script for those who want it to display only on the primary monitor. 
Images should load slight faster too, so that should help on older/slower computers (though this isn't a problem for a computer with a decent CPU).

Also slightly easier to read/understand as there's much less going on.